requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
Pillow==10.0.1
Werkzeug==2.3.7
gunicorn==21.2.0